# docs/openapi.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: MoneyMinder - Rest api
  description: "Api for MoneyMinder project"
...
paths:
...
  /users/register:
    post:
      tags:
        - users
      description: Register user
      operationId: registerUser
      requestBody:
        description: User details to register new user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequestBody'
        required: true
      responses:
        '201':
          description: Successfully registered user
        '400':
          description: Bad request
        '409':
          description: User already exists
  /users/login:
    post:
      tags:
        - users
      description: login user
      operationId: loginUser
      requestBody:
        description: User details to login to the account
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginUserRequestBody'
        required: true
      responses:
        '200':
          description: logged in successfully
          headers:
            JsonWebToken:
              $ref: '#/components/headers/JsonWebToken'
        '400':
          description: Bad request
        '401':
          description: user is not verified
        '404':
          description: User with that email does not exists
        '403':
          description: User couldn't be authenticated
  /users/change-password:
    put:
      tags:
        - users
      description: change password for user
      operationId: changePassword
      requestBody:
        description: new password
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequestBody'
        required: true
      responses:
        '200':
          description: changed password successfully
          headers:
            JsonWebToken:
              $ref: '#/components/headers/JsonWebToken'
        '400':
          description: Bad request
        '404':
          description: User not found
        '403':
          description: User couldn't be authenticated
  /users/reset-password:
    post:
      tags:
        - users
      description: reset password for user
      operationId: resetPassword
      requestBody:
        description: reset password
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequestBody'
        required: true
      responses:
        '200':
          description: successfully send message with link to reset password
        '400':
          description: Bad request
        '404':
          description: User not found
  /users/confirm-reset-password:
    put:
      tags:
        - users
      description: set a new password for user
      operationId: confirmResetPassword
      requestBody:
        description: new password
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmResetPasswordRequestBody'
        required: true
      parameters:
        - name: ResetPasswordToken
          in: query
          description: Reset password token
          required: true
          schema:
            type: string
            example: '62eiredl-ORKa-4930-9831-eec546b91830'
      responses:
        '200':
          description: successfully set a new password for user
        '400':
          description: password is invalid
        '404':
          description: Token is expired or does not exist
...
components:
  schemas:
...
    CreateUserRequestBody:
      required:
        - name
        - email
        - password
      type: object
      properties:
        name:
          type: string
          example: 'verion'

        password:
          type: string
          example: 'K#lasdf&82z'
        email:
          type: string
          example: 'verion@gmail.com'
    LoginUserRequestBody:
      required:
        - email
        - password
      type: object
      properties:
        email:
          type: string
          example: 'verion@gmail.com'
        password:
          type: string
          example: 'K#lasdf&82z'
    ChangePasswordRequestBody:
      required:
        - oldPassword
        - repeatOldPassword
        - newPassword
      type: object
      properties:
        oldPassword:
          type: string
          example: '123456'
        repeatOldPassword:
          type: string
          example: '123456'
        newPassword:
          type: string
          example: 'qwerty'
    ResetPasswordRequestBody:
      required:
        - email
      type: object
      properties:
        email:
          type: string
          example: 'verion@gmail.com'
    ConfirmResetPasswordRequestBody:
      required:
        - new password
      type: object
      properties:
        password:
          type: string
          example: '1234567'
  headers:
    JsonWebToken:
      schema:
        description: JsonWebToken
        type: string
        format: application/jwt
        pattern: ^([a-zA-Z0-9_=]+)\.([a-zA-Z0-9_=]+)\.([a-zA-Z0-9_\-\+\/=]+)$
        example: "eyJhbGciOiJIUzI1NiJ9
        .eyJzdWIiOiJ0ZXN0b3d5Lm1haWxAZ21haWwuY29tIiwiaWF0IjoxN
        zAzODk1OTg0LCJleHAiOjE3MDM5MTc1ODR9
        .Ubh9XvZvp9qbu624AtiRBPXj1BzscGVC1MyqfHL5s60"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT